DynamicAutodiff

Documentation for DynamicAutodiff.

DynamicAutodiff.EvaluatesToConstant

DynamicAutodiff.D

DynamicAutodiff.EvaluatesToConstant — Type

EvaluatesToConstant

Used to declare if an operator will always evaluate to a constant.

DynamicAutodiff.D — Method

D(ex::AbstractExpression, feature::Integer)

Compute the derivative of ex with respect to the feature-th variable. Returns a new expression with an expanded set of operators.